Milestone 10 code added

+++ assignment1.py
@@ -10,6 +10,7 @@ import re
 import json
 import csv
 
+import os
 
 # Fetching the Movie ids for top 5 movies
 def movies_id_func():
@@ -30,30 +31,37 @@ def movies_id_func():
         print("Invalid URL")
 
 ids = movies_id_func()
-
 # Fetching the synposis for the above movies and storings
 
 def movies_synopsis_func(movie_id):
     movies_synopsis = []
 
-    for i in movie_id:
-        request_movie_link = requests.get(f"https://www.imdb.com/title/{i}")
-        if request_movie_link.status_code == 200:
-            format_response = bs4.BeautifulSoup(request_movie_link.text, 'lxml')            
-            movies_synopsis.append(format_response.select('.ipc-html-content')[1].getText())
-        else:
-            return "Incorrect URL"
-    return movies_synopsis
+    try:
+        for i in movie_id:
+            request_movie_link = requests.get(f"https://www.imdb.com/title/{i}")
+            if request_movie_link.status_code == 200:
+                format_response = bs4.BeautifulSoup(request_movie_link.text, 'lxml')            
+                movies_synopsis.append(format_response.select('.ipc-html-content')[1].getText())
+            else:
+                print("Incorrect URL")
+        return movies_synopsis
+    except TypeError:
+        print("None object returned")
+    except IndexError:
+        print("Index out of range")
 synopsis = movies_synopsis_func(ids)
 
 # creating a bag of words of synopsis and addin that in the dictionary with key as film_id
 
 def bag_of_words(string):
-    stop_words = set(stopwords.words('english'))
-    word_tokens = word_tokenize(string)
-    filtered_sentence = [word for word in word_tokens if not word.lower() in stop_words]
-    bag_of_word = ' '.join(filtered_sentence)
-    return bag_of_word
+    try:
+        stop_words = set(stopwords.words('english'))
+        word_tokens = word_tokenize(string)
+        filtered_sentence = [word for word in word_tokens if not word.lower() in stop_words]
+        bag_of_word = ' '.join(filtered_sentence)
+        return bag_of_word
+    except TypeError:
+        print("Missing argument")
 
 dictionary_movie_data = {}
 i = 0
@@ -67,24 +75,52 @@ for string in synopsis:
 # api data fetching
 
 def fetch_api_data(id):
-    pattern = r"\D{2}\d{7}"
-    if re.compile(pattern).match(id).group()==id:
-        response = requests.get(f"http://www.omdbapi.com/?i={id}&apikey=1db04143")
-        api_info = response.json()
-        return api_info
-
+    try:
+        pattern = r"\D{2}\d{7}"
+        if re.compile(pattern).match(id).group()==id:
+            response = requests.get(f"http://www.omdbapi.com/?i={id}&apikey=1db04143")
+            api_info = response.json()
+            return api_info
+    except TypeError:
+        print("Missing arguments")
 
 for i in ids:
     api_call = fetch_api_data(i)
     dictionary_movie_data[i]['Genre'] = api_call['Genre']
     dictionary_movie_data[i]['Actors'] = api_call['Actors']
 
-
 fields = ['movie_id', 'Synopsis', 'Genre', 'Actors']
 
-with open("movies_data.csv", "w") as csvfile:
-    writer = csv.DictWriter(csvfile, fields)
-    writer.writeheader()
-    for key in dictionary_movie_data:
-        writer.writerow({field: dictionary_movie_data[key].get(field) or key for field in fields})
+def write_csv():
+    if os.path.exists('movies_data.csv'):
+        csvfile = open("movies_data.csv", "r")
+        reader = csv.DictReader(csvfile)
+        for row in reader:
+            for item in ids:
+                if row['movie_id']==item:
+                    dictionary_movie_data.pop(item)
 
+        csvfile.close()
+        if len(dictionary_movie_data)>0:
+            with open("movies_data.csv", "a") as file:
+                writer = csv.DictWriter(file, fields)
+                for key in dictionary_movie_data:
+                    writer.writerow({field: dictionary_movie_data[key].get(field) or key for field in fields})
+            
+    else:
+        with open("movies_data.csv", "w") as csvfile:
+            writer = csv.DictWriter(csvfile, fields)
+            writer.writeheader()
+            for key in dictionary_movie_data:
+                writer.writerow({field: dictionary_movie_data[key].get(field) or key for field in fields})
+
+write_csv()
+
+def fetch_movies_data():
+    item = input("Enter Genre or Actor name for data fetching: ")
+    with open("movies_data.csv", 'r') as file:
+        reader = csv.DictReader(file)
+        rows = [row for row in reader if row['Actors']==item or row['Genre']==item]
+        return rows
+
+print(fetch_movies_data())
